Add files via upload
Added initial class and tests.
The tests are not all-encompassing, or using a unit-testing framework, nor are they entirely automated. They produce text output to stdout, for you to look at, and see that the random distribution "loooks about right", and in this way it is probably more aptly called "example usage code".